requirements.txt
pytorch==1.5.0
torchvision==0.7.0
cudatoolkit=10.1